Hopefully fixed register lookup bug.

+++ Signals.cs
@@ -99,7 +99,7 @@ namespace mu2e.FEB_Test_Jig
             Biases[1] = new BiasSignal();
         }
 
-        public int myFPGA_ID { get { return Biases[0].myFPGA_ID; } }
+        public ushort myFPGA_ID { get { return Biases[0].myFPGA_ID; } }
         public int signalIndex { get { return Biases[0].signalIndex; } }
         public Mu2e_Register register { get { return Biases[0].register; } }
